+++ ReleaseNotes.md
@@ -1,5 +1,17 @@
 # Release Notes SBBMobileDesignSystemSwiftUI für iOS & SwiftUI
 
+## V0.0.6 - 24 Feb 2020
+
+### Features
+* SBBSegmentedPicker
+* SBBDivider
+* new semantic SBBColor .controlBackground
+* new SBBFont .tableHeader
+
+### Bugfixes
+* SBBBubbleView background red now reflects the opacity from the SBBNavigationBar (to be removed once the opacity can be set to false on the NavBar)
+* SBBBubbleView voiceover now only reads the hint, if the view is expandable (i.e. has a detail text)
+
 ## V0.0.5 - 14 Feb 2020
 
 ### Features